+++ src/ft_strncmp.c
@@ -6,19 +6,20 @@
 /*   By: ydidenko <ydidenko@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/06 13:32:52 by ydidenko      #+#    #+#                 */
-/*   Updated: 2023/10/17 19:35:11 by ydidenko      ########   odam.nl         */
+/*   Updated: 2023/10/18 11:18:11 by ydidenko      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stddef.h>
 /*
 ** The strncmp() function compares not more than n characters.
 ** Because strncmp() is designed for comparing strings rather than binary data,
 ** characters that appear after a `\0' character are not compared.
 */
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
